Refactor WriteFreelyModel extensions into separate files
